requirements.txt
requests~=2.32.3
flask~=3.0.3
gunicorn~=23.0.0
flask-caching~=2.2.0
python-dotenv~=1.0.1